Refactors parset into a separate file.
