+++ daliuge-engine/dlg/graph_loader.py
@@ -42,9 +42,9 @@ from .drop import (
     NullDROP,
     EndDROP,
     PlasmaDROP,
-    PlasmaFlightDROP,
-    ParameterSetDROP
+    PlasmaFlightDROP
 )
+from dlg.parset_drop import ParameterSetDROP
 from .exceptions import InvalidGraphException
 from .json_drop import JsonDROP
 from .common import Categories, DropType
